# update nodegit usage in integration test to only clone test repo once