--- a/.vscode/settings.json
+++ b/.vscode/settings.json
@@ -1,6 +1,6 @@
 {
     "solidity.compileUsingRemoteVersion": "latest",
     "solidity.packageDefaultDependenciesContractsDirectory": "contracts",
-    "mochaExplorer.files": "integration/**/*.ts",
+    "mochaExplorer.files": "integration/**/*.test.ts",
     "mochaExplorer.require": "ts-node/register"
 }