@DmitriyShepelev 2022-10-23
Since the tests are now compatible in `--parallel` mode, should we add `--parallel` to the `test` script in the `package.json` file?