@mattstam 2022-10-23
> Since the tests are now compatible in `--parallel` mode, should we add `--parallel` to the `test` script in the `package.json` file?

This would make it faster to run in the normal case, but for typical developer usage there is one problem with this: you can't run a single mocha test `it.only()` when the `--parallel` flag is on. This use-case is common enough when adding functionality that it doesn't seem worth it to throw a possible headache at people trying to add a new test for their contract change.

I could see the argument for adding more package.json commands (e.g. `npm run testp` to run with `--parallel`), but at a certain point you're just going to want to start using the raw `npx hardhat` to test with additional flags instead so you don't have a bloated `package.json` with a bunch of unknown, project-specific aliases.